+++ src/components/Footer.js
@@ -8,7 +8,13 @@ export default function Footer() {
           <Link to="/">Link</Link>
           <Link to="/">Link</Link>
         </div>
-        <span>logo</span>
+        <a href="/">
+          <img
+            className="max-h-10"
+            src={window.matchMedia('(prefers-color-scheme: dark)').matches ? '/svg/logo-dark.svg' : '/svg/logo.svg'}
+            alt="logo"
+          />
+        </a>
         <div className="space-x-4">
           <Link to="/">Link</Link>
           <Link to="/">Link</Link>